# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stackr API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/domain-resources": {
      "get": {
        "summary": "List domain resources (scope: stacks:read)",
        "operationId": "listDomainResources",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1DomainResourceOut"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "post": {
        "summary": "Create domain resource (scope: domains:write)",
        "operationId": "createDomainResource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1DomainResourceIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1DomainResourceOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/orgs": {
      "get": {
        "summary": "List organizations (scope: stacks:read)",
        "operationId": "listOrganizations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1OrgOut"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/orgs/{id}/config/plan": {
      "post": {
        "summary": "Plan org config (scope: config:apply)",
        "operationId": "planOrgConfig",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1PlanOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/orgs/{id}/settings": {
      "get": {
        "summary": "Get organization defaults (scope: stacks:read)",
        "operationId": "getOrganizationDefaults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1SettingsOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "patch": {
        "summary": "Set organization defaults (scope: stacks:write)",
        "operationId": "setOrganizationDefaults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1SettingsOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "summary": "Get server defaults (scope: stacks:read)",
        "operationId": "getServerDefaults",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1SettingsOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "patch": {
        "summary": "Set server defaults (scope: stacks:write)",
        "operationId": "setServerDefaults",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1SettingsOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "V1DomainResourceIn": {
        "required": [
          "host"
        ],
        "type": "object",
        "properties": {
          "acme_email": {
            "type": "string",
            "description": "Let's Encrypt account for certificates under this host; empty uses the instance's"
          },
          "host": {
            "type": "string",
            "description": "bare lowercase hostname, e.g. example.com"
          },
          "include_env_on_default": {
            "type": "boolean",
            "description": "keep the default env's slug in generated names"
          },
          "level": {
            "type": "string",
            "description": "instance or node (default, admins only) | org | stack"
          },
          "owner": {
            "type": "string",
            "description": "org or stack id the resource belongs to; ignored for instance level"
          }
        }
      },
      "V1DomainResourceOut": {
        "type": "object",
        "properties": {
          "acme_email": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "include_env_on_default": {
            "type": "boolean"
          },
          "level": {
            "type": "string"
          },
          "owner_id": {
            "type": "string"
          }
        }
      },
...
      "V1ErrBody": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/V1ErrDetail"
          }
        }
      },
      "V1ErrDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "V1OrgOut": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "V1PlanOut": {
        "type": "object",
        "properties": {
          "commit_sha": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "decided_at": {
            "type": "string"
          },
          "destructive": {
            "type": "boolean"
          },
          "env_slug": {
            "type": "string",
            "description": "empty for a stack-scoped plan"
          },
          "error": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "stack_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "pending | clean | applied | rejected | superseded | error"
          },
          "summary": {
            "type": "string"
          }
        }
      },
...
      "V1SettingKnob": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "own": {
            "type": "string",
            "description": "this level's own override; absent means inherited",
            "nullable": true
          },
          "source": {
            "type": "string",
            "description": "server | org | stack | env | built-in"
          },
          "value": {
            "type": "string",
            "description": "the resolved value at this level"
          }
        }
      },
      "V1SettingsOut": {
        "type": "object",
        "properties": {
          "level": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V1SettingKnob"
            },
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "API key"
      }
    }
  }
}